[Refactor] Journey -> Travel 명칭 변경
