--- app/src/main/java/com/example/moiz/domain/model/Journey.kt
@@ -1,8 +1,0 @@
-package com.example.moiz.domain.model
-
-data class Journey(
-    val title: String,
-    val startDate: String,
-    val endDate: String,
-    val memberList: ArrayList<String>,
-) : java.io.Serializable
