+++ app/src/main/res/layout/create_travel_list_fragment.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="vm"
-            type="com.example.moiz.presentation.createJourneyList.CreateJourneyListViewModel" />
+            type="com.example.moiz.presentation.createTravelList.CreateTravelListViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
