+++ app/src/main/res/navigation/main_navigation.xml
@@ -11,12 +11,12 @@
         android:label="MainFragment"
         tools:layout="@layout/main_fragment">
         <action
-            android:id="@+id/action_mainFragment_to_createJourneyListFragment"
-            app:destination="@id/createJourneyListFragment" />
+            android:id="@+id/action_mainFragment_to_createTravelListFragment"
+            app:destination="@id/createTravelListFragment" />
     </fragment>
     <fragment
-        android:id="@+id/createJourneyListFragment"
-        android:name="com.example.moiz.presentation.createJourneyList.CreateJourneyListFragment"
-        android:label="CreateJourneyListFragment"
-        tools:layout="@layout/create_journey_list_fragment" />
+        android:id="@+id/createTravelListFragment"
+        android:name="com.example.moiz.presentation.createTravelList.CreateTravelListFragment"
+        android:label="CreateTravelListFragment"
+        tools:layout="@layout/create_travel_list_fragment" />
 </navigation>
